WIP: Use persistent browser context with profiles
